+++ CHANGELOG.md
@@ -1,3 +1,16 @@
-## [0.1.0-alpha] - 2021-04-30
+# Flutter Scroll Shadow
+## [0.1.0-alpha.2] - 2021-04-30
+#### [@Zabadam](https://github.com/zabadam)
+- Added `Curve` parameter and made `duration` a full `Duration`.
 
-* Initial Open Source release.
+## [0.1.0-alpha.1] - 2021-04-30
+#### [@Zabadam](https://github.com/zabadam)
+- Split package into `/src` files.
+- Null safety support.
+- Default `duration` increased to `300` ms (and property name shortened).
+- Major simplification of visibility logic.
+- Removal of `AnimationController`s and `Tween`s in favor of `AnimatedOpacity`s.
+- `loc`: 231 -> 185, with added comments 👍
+
+## [0.1.0-alpha] - 2021-04-30
+- Initial release.